requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
boto3==1.34.145
botocore==1.34.145
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
distro==1.9.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.6
django-allauth==0.63.3
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
durationpy==0.7
exceptiongroup==1.1.3
frozenlist==1.4.1
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.34.1
importlib_metadata==8.4.0
inflection==0.5.1
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pillow==10.3.0
placebo==0.9.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.5.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
trio==0.22.2
trio-websocket==0.10.4
troposphere==4.8.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.0
asttokens==2.4.1
django-environ==0.11.2
django-extensions==3.2.3
executing==2.0.1
geographiclib==2.0
geopy==2.4.1
ipython==8.13.0
jedi==0.19.1
joblib==1.4.2
matplotlib-inline==0.1.7
numpy==1.24.4
parso==0.8.4
prompt_toolkit==3.0.47
pure-eval==0.2.2
Pygments==2.18.0
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
wcwidth==0.2.13